edugen-agent/requirements.txt
streamlit
requests
gtts
python-dotenv
